add --hard flag to rollback block as well as state

+++ state/rollback.go
@@ -12,7 +12,7 @@ import (
 // Rollback overwrites the current Tendermint state (height n) with the most
 // recent previous state (height n - 1).
 // Note that this function does not affect application state.
-func Rollback(bs BlockStore, ss Store) (int64, []byte, error) {
+func Rollback(bs BlockStore, ss Store, removeBlock bool) (int64, []byte, error) {
 	invalidState, err := ss.Load()
 	if err != nil {
 		return -1, nil, err
@@ -27,11 +27,12 @@ func Rollback(bs BlockStore, ss Store) (int64, []byte, error) {
 	// when the user stopped the node the state wasn't updated but the blockstore was. Discard the
 	// pending block before continuing.
 	if height == invalidState.LastBlockHeight+1 {
-		if err := bs.DeleteLatestBlock(); err != nil {
-			return -1, nil, fmt.Errorf("failed to save rolled back blockstore: %w", err)
+		if removeBlock {
+			if err := bs.DeleteLatestBlock(); err != nil {
+				return -1, nil, fmt.Errorf("failed to remove final block from blockstore: %w", err)
+			}
 		}
-
-		height = bs.Height()
+		return invalidState.LastBlockHeight, invalidState.AppHash, nil
 	}
 
 	// If the state store isn't one below nor equal to the blockstore height than this violates the
@@ -112,9 +113,12 @@ func Rollback(bs BlockStore, ss Store) (int64, []byte, error) {
 		return -1, nil, fmt.Errorf("failed to save rolled back state: %w", err)
 	}
 
-	// remove the block associated with the previous state.
-	if err := bs.DeleteLatestBlock(); err != nil {
-		return -1, nil, fmt.Errorf("failed to save rolled back blockstore: %w", err)
+	// If removeBlock is true then also remove the block associated with the previous state.
+	// This will mean both the last state and last block height is equal to n - 1
+	if removeBlock {
+		if err := bs.DeleteLatestBlock(); err != nil {
+			return -1, nil, fmt.Errorf("failed to remove final block from blockstore: %w", err)
+		}
 	}
 
 	return rolledBackState.LastBlockHeight, rolledBackState.AppHash, nil
